django/requirements.txt
-i https://pypi.org/simple
arabic-reshaper==2.1.3
asgiref==3.3.4; python_version >= '3.6'
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-rest-framework-nested==0.0.1
django==3.2.2
djangorestframework==3.12.4
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gitdb==4.0.7; python_version >= '3.4'
gitpython==3.1.14
gunicorn==20.1.0
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
kiwisolver==1.3.1; python_version >= '3.6'
markdown==3.3.4
matplotlib==3.4.2
numpy==1.20.3; python_version >= '3.7'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.2.4
pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydot==1.4.2
pyjwt[crypto]==2.1.0; python_version >= '3.6'
pyparsing==2.4.7
pypdf2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.67; python_version >= '2.7' and python_version < '4'
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '3.0'
scipy==1.6.3; python_version < '3.10' and python_version >= '3.7'
seaborn==0.11.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==4.0.0; python_version >= '3.5'
sqlparse==0.4.1; python_version >= '3.5'
stripe==2.56.0
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
